pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "echograph"
version = "0.1.0"
description = "Document ingestion and relationship discovery toolkit for cloud guidelines"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.10"
authors = [{ name = "EchoGraph Maintainers" }]
dependencies = [
  "pdfplumber>=0.10.4",
  "python-docx>=1.0.0",
  "tika>=2.6.0",
  "requests>=2.31.0",
  "pydantic>=2.5.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "numpy>=1.24.0",
  "sentence-transformers>=2.2.2",
  "qdrant-client>=1.7.3",
  "sqlalchemy>=2.0.0",
  "psycopg[binary,pool]>=3.1.12",
  "asyncpg>=0.29.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.23.0",
  "httpx>=0.25.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.1",
  "ruff>=0.1.0",
  "mypy>=1.6.1",
  "types-requests",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
markers = [
  "integration: marks tests that hit databases or external services",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.coverage.run]
source = ["ingestion", "processing", "api"]
